requirements.txt
aioredis==1.3.1
altgraph==0.17
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
bleach==3.1.5
bottle==0.12.18
bottle-websocket==0.2.9
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
cirq==0.8.0
click==7.1.2
colab-ssh==0.2.60
constantly==15.1.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.19
daphne==2.5.0
decorator==4.4.2
dill==0.3.1.1
distro==1.5.0
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-ipware==3.0.1
django-xff==1.3.0
docplex==2.15.194
Eel==0.12.3
fastjsonschema==2.14.4
freezegun==0.3.15
future==0.18.2
gevent==20.5.0
gevent-websocket==0.10.1
google-api-core==1.17.0
google-auth==1.15.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
hiredis==1.1.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
joblib==0.14.1
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.2.0
marshmallow==3.6.0
marshmallow-polyfield==5.9
matplotlib==3.2.1
MouseInfo==0.1.3
mpmath==1.1.0
msgpack==1.0.0
nest-asyncio==1.3.3
networkx==2.4
nltk==3.5
ntlm-auth==1.4.0
numpy==1.18.4
opencv-python==4.2.0.34
ordered-set==4.0.1
packaging==20.4
pandas==1.0.5
pefile==2019.4.18
Pillow==7.1.2
ply==3.11
protobuf==3.8.0
psutil==5.7.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pybind11==2.5.0
pycparser==2.20
PyGetWindow==0.0.8
PyHamcrest==2.0.2
PyInstaller==3.6
PyLaTeX==1.3.2
PyMsgBox==1.0.7
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
python-constraint==1.4.0
python-dateutil==2.8.1
python-decouple==3.3
PyTweening==1.0.3
sqlparse==0.3.1

